Server/requirements.txt
fastmcp==2.12.0
fastmcp[cli]
uvicorn==0.35.0
uv==0.8.14

requests==2.32.5

# Telemetry (optional but recommended)
posthog==3.7.0

